NetSNMP v0.0.30 NetSNMP

A Net-SNMP library supporting SNMPv1/2c/3.

Functions

credential(args)

Specs

credential(list) :: Keyword.t

Returns a keyword list containing the given SNMPv1/2c/3 credentials.

Examples

iex> NetSNMP.credential [:v1, "public"]
[version: "1", community: "public"]

iex> NetSNMP.credential [:v2c, "public"]
[version: "2c", community: "public"]

iex> NetSNMP.credential [:v3, :no_auth_no_priv, "user"]
[version: "3", sec_level: "noAuthNoPriv", sec_name: "user"]

iex> NetSNMP.credential [:v3, :auth_no_priv, "user", :sha, "authpass"]
[ version: "3",
  sec_level: "authNoPriv",
  sec_name: "user",
  auth_proto: "sha", auth_pass: "authpass"
]

iex> NetSNMP.credential [:v3, :auth_priv, "user", :sha, "authpass", :aes, "privpass"]
[ version: "3",
  sec_level: "authPriv",
  sec_name: "user",
  auth_proto: "sha", auth_pass: "authpass",
  priv_proto: "aes", priv_pass: "privpass"
]

credential(version, community)

Specs

credential(:v1, String.t) :: Keyword.t

credential(:v2c, String.t) :: Keyword.t

Returns a keyword list containing the given SNMPv1/2c community.

Examples

iex> NetSNMP.credential :v1, "public"
[version: "1", community: "public"]

iex> NetSNMP.credential :v2c, "public"
[version: "2c", community: "public"]

credential(version, sec_level, sec_name)

Specs

credential(:v3, :no_auth_no_priv, String.t) :: Keyword.t

Returns a keyword list containing the given SNMPv3 noAuthNoPriv credentials.

Examples

iex> NetSNMP.credential :v3, :no_auth_no_priv, "user"
[version: "3", sec_level: "noAuthNoPriv", sec_name: "user"]

credential(version, sec_level, sec_name, auth_proto, auth_pass)

Specs

credential(:v3, :auth_no_priv, String.t, :md5 | :sha, String.t) :: Keyword.t

Returns a keyword list containing the given SNMPv3 authNoPriv credentials.

Examples

iex> NetSNMP.credential :v3, :auth_no_priv, "user", :sha, "authpass"
[ version: "3",
  sec_level: "authNoPriv",
  sec_name: "user",
  auth_proto: "sha", auth_pass: "authpass"
]

credential(version, sec_level, sec_name, auth_proto, auth_pass, priv_proto, priv_pass)

Specs

credential(:v3, :auth_priv, String.t, :md5 | :sha, String.t, :des | :aes, String.t) :: Keyword.t

Returns a keyword list containing the given SNMPv3 authPriv credentials.

Examples

iex> NetSNMP.credential :v3, :auth_priv, "user", :sha, "authpass", :aes, "privpass"
[ version: "3",
  sec_level: "authPriv",
  sec_name: "user",
  auth_proto: "sha", auth_pass: "authpass",
  priv_proto: "aes", priv_pass: "privpass"
]

get(snmp_object, uri, credential, context \\ "")

Send SNMPGET request to uri for given objects and credentials.

set(snmp_object, uri, credential, context \\ "")

Send SNMPSET request to uri for given objects and credentials.

table(snmp_object, uri, credential, context \\ "")

Perform SNMP table operations against uri for given objects and credentials.

walk(snmp_object, uri, credential, context \\ "")

Send SNMPGETNEXT requests to uri, starting at given objects, with given credentials.